refactor: split webauthn device info row into a separate component
