+++ web/src/views/Settings/TwoFactorAuthSettings.tsx
@@ -32,6 +32,8 @@ import { getWebauthnDevices } from "@root/services/UserWebauthnDevices";
 import { initiateWebauthnRegistrationProcess } from "@services/RegisterDevice";
 import { AutheliaState, AuthenticationLevel } from "@services/State";
 
+import WebauthnDeviceItem from "./WebauthnDeviceItem";
+
 interface Props {
     state: AutheliaState;
 }
@@ -110,154 +112,12 @@ export default function TwoFactorAuthSettings(props: Props) {
                             {webauthnDevices
                                 ? webauthnDevices.map((x, idx) => {
                                       return (
-                                          <React.Fragment>
-                                              <TableRow
-                                                  sx={{ "& > *": { borderBottom: "unset" } }}
-                                                  key={x.kid.toString()}
-                                              >
-                                                  <TableCell>
-                                                      <Tooltip title={translate("Show Details")} placement="right">
-                                                          <IconButton
-                                                              aria-label="expand row"
-                                                              size="small"
-                                                              onClick={() => handleWebAuthnDetailsChange(idx)}
-                                                          >
-                                                              {webauthnShowDetails === idx ? (
-                                                                  <KeyboardArrowUpIcon />
-                                                              ) : (
-                                                                  <KeyboardArrowDownIcon />
-                                                              )}
-                                                          </IconButton>
-                                                      </Tooltip>
-                                                  </TableCell>
-                                                  <TableCell component="th" scope="row">
-                                                      {x.description}
-                                                  </TableCell>
-                                                  <TableCell>
-                                                      <Switch defaultChecked={false} size="small" />
-                                                  </TableCell>
-                                                  <TableCell align="center">
-                                                      <Stack
-                                                          direction="row"
-                                                          spacing={1}
-                                                          alignItems="center"
-                                                          justifyContent="center"
-                                                      >
-                                                          <Tooltip title={translate("Edit")} placement="bottom">
-                                                              <IconButton aria-label="edit">
-                                                                  <EditIcon />
-                                                              </IconButton>
-                                                          </Tooltip>
-                                                          <Tooltip title={translate("Delete")} placement="bottom">
-                                                              <IconButton aria-label="delete">
-                                                                  <DeleteIcon />
-                                                              </IconButton>
-                                                          </Tooltip>
-                                                      </Stack>
-                                                  </TableCell>
-                                              </TableRow>
-                                              <TableRow>
-                                                  <TableCell style={{ paddingBottom: 0, paddingTop: 0 }} colSpan={4}>
-                                                      <Collapse
-                                                          in={webauthnShowDetails === idx}
-                                                          timeout="auto"
-                                                          unmountOnExit
-                                                      >
-                                                          <Grid container spacing={2} sx={{ mb: 3, margin: 1 }}>
-                                                              <Grid item xs={12} sm={12} md={12} lg={12} xl={12}>
-                                                                  <Box sx={{ margin: 1 }}>
-                                                                      <Typography
-                                                                          variant="h6"
-                                                                          gutterBottom
-                                                                          component="div"
-                                                                      >
-                                                                          {translate("Details")}
-                                                                      </Typography>
-                                                                  </Box>
-                                                              </Grid>
-                                                              <Grid item xs={12} sm={12} md={12} lg={12} xl={12}>
-                                                                  <Divider variant="middle" />
-                                                              </Grid>
-                                                              <Grid item xs={12} sm={12} md={12} lg={12} xl={12}>
-                                                                  <Typography>
-                                                                      {translate("Webauthn Credential Identifier", {
-                                                                          id: x.kid.toString(),
-                                                                      })}
-                                                                  </Typography>
-                                                              </Grid>
-                                                              <Grid item xs={12} sm={12} md={12} lg={12} xl={12}>
-                                                                  <Typography>
-                                                                      Public Key: {x.public_key}
-                                                                      {translate("Webauthn Public Key", {
-                                                                          key: x.public_key.toString(),
-                                                                      })}
-                                                                  </Typography>
-                                                              </Grid>
-                                                              <Grid item xs={12} sm={12} md={12} lg={12} xl={12}>
-                                                                  <Divider variant="middle" />
-                                                              </Grid>
-                                                              <Grid item xs={6} sm={6} md={4} lg={4} xl={3}>
-                                                                  <Typography>
-                                                                      {translate("Relying Party ID")}
-                                                                  </Typography>
-                                                                  <Typography>{x.rpid}</Typography>
-                                                              </Grid>
-                                                              <Grid item xs={6} sm={6} md={4} lg={4} xl={3}>
-                                                                  <Typography>
-                                                                      {translate("Authenticator Attestation GUID")}
-                                                                  </Typography>
-                                                                  <Typography>{x.aaguid}</Typography>
-                                                              </Grid>
-                                                              <Grid item xs={6} sm={6} md={4} lg={4} xl={3}>
-                                                                  <Typography>
-                                                                      {translate("Attestation Type")}
-                                                                  </Typography>
-                                                                  <Typography>{x.attestation_type}</Typography>
-                                                              </Grid>
-                                                              <Grid item xs={6} sm={6} md={4} lg={4} xl={3}>
-                                                                  <Typography>{translate("Transports")}</Typography>
-                                                                  <Typography>
-                                                                      {x.transports.length === 0
-                                                                          ? "N/A"
-                                                                          : x.transports.join(", ")}
-                                                                  </Typography>
-                                                              </Grid>
-                                                              <Grid item xs={6} sm={6} md={4} lg={4} xl={3}>
-                                                                  <Typography>{translate("Clone Warning")}</Typography>
-                                                                  <Typography>
-                                                                      {x.clone_warning
-                                                                          ? translate("Yes")
-                                                                          : translate("No")}
-                                                                  </Typography>
-                                                              </Grid>
-                                                              <Grid item xs={6} sm={6} md={4} lg={4} xl={3}>
-                                                                  <Typography>{translate("Created")}</Typography>
-                                                                  <Typography>{x.created_at.toString()}</Typography>
-                                                              </Grid>
-                                                              <Grid item xs={6} sm={6} md={4} lg={4} xl={3}>
-                                                                  <Typography>{translate("Last Used")}</Typography>
-                                                                  <Typography>
-                                                                      {x.last_used_at === undefined
-                                                                          ? translate("Never")
-                                                                          : x.last_used_at.toString()}
-                                                                  </Typography>
-                                                              </Grid>
-                                                              <Grid item xs={6} sm={6} md={4} lg={4} xl={3}>
-                                                                  <Typography>{translate("Usage Count")}</Typography>
-                                                                  <Typography>
-                                                                      {x.sign_count === 0
-                                                                          ? translate("Never")
-                                                                          : x.sign_count}
-                                                                  </Typography>
-                                                              </Grid>
-                                                              <Grid item xs={12} sm={12} md={12} lg={12} xl={12}>
-                                                                  <Divider variant="middle" />
-                                                              </Grid>
-                                                          </Grid>
-                                                      </Collapse>
-                                                  </TableCell>
-                                              </TableRow>
-                                          </React.Fragment>
+                                          <WebauthnDeviceItem
+                                              device={x}
+                                              idx={idx}
+                                              webauthnShowDetails={webauthnShowDetails}
+                                              handleWebAuthnDetailsChange={handleWebAuthnDetailsChange}
+                                          />
                                       );
                                   })
                                 : null}
